Modify to make app use Material components
- Also created new SelectableLists/ListsItem components

+++ src/pages/Main.js
@@ -1,37 +1,42 @@
 import { useState } from 'react'
-import TagsSidebar from 'components/TagsSidebar';
-import ListsSidebar from 'components/ListsSidebar';
-import Task from 'components/Task'
+
 import ResponsivePage from "components/ResponsivePage"
 import Header from "components/Header"
-
+import ListsSidebar from 'components/ListsSidebar';
+import TagsSidebar from 'components/TagsSidebar';
+import Task from 'components/Task'
+import Button from 'material/Button'
 
 import 'pages/Main.css'
-
 
 const Main = ({ context }) => {
 
   const tasks = context.getTasks()
-
   const [currentList, setCurrentList] = useState(null)
 
   context.setCurrentListCallbacks(() => currentList, setCurrentList)
 
-
   const editListClicked = (e) => {
-    context.editList(currentList, {
-      "text": prompt("Please enter name of list")
-    })
-
+    const userInput = prompt("Please enter name of list")
+    if (userInput) {
+      context.editList(currentList, {
+        "text": userInput
+      })
+    }
   }
+
   const deleteListClicked = (e) => {
+    const prompt = `You are abou to delete this list permenantly. Continue?`
+    if (!window.confirm(prompt)) {
+      return
+    }
     context.deleteList(currentList)
     setCurrentList(null)
     // Also delete tasks locally?
   }
+  const currentListName = currentList ? context.getLists().find((list) => list.id === currentList).text : null
 
   return (
-
     <ResponsivePage
       header={<Header context={context} />}
       drawer={<ListsSidebar context={context} />}
@@ -51,9 +56,13 @@ const Main = ({ context }) => {
               ?
               <>
                 <div>
-                  <span>{context.getLists().find((list) => list.id === currentList).text}</span>
-                  <button onClick={editListClicked}>Edit name</button>
-                  <button onClick={deleteListClicked}>Delete</button>
+                  <span>{currentListName}</span>
+                  <Button onClick={editListClicked}>
+                    Edit name
+                  </Button>
+                  <Button onClick={deleteListClicked}>
+                    Delete
+                  </Button>
                 </div>
                 <div id="task-container">
                   {
